+++ lib/ace/http2/server.ex
@@ -1,4 +1,35 @@
 defmodule Ace.HTTP2.Server do
+  @moduledoc """
+  Server to handle an the HTTP/2.0 connection of a single client
+
+  # Needs docs on stream handlers
+
+  #### Example
+
+  ```elixir
+  defmodule MyApp.StreamHandler do
+    use GenServer
+
+    def start_link(config) do
+      GenServer.start_link(__MODULE__, config)
+    end
+
+    def handle_info({stream, {:headers, _}}, state) do
+      response_headers = %{
+        headers: [{":status", "200"}, {"content-length", "13"}],
+        end_stream: false
+      }
+      Server.send(stream, response_headers)
+      response_body = %{
+        data: "Hello, World!",
+        end_stream: true
+      }
+      Server.send(stream, response_body)
+      {:stop, :normal, state}
+    end
+  end
+  ```
+  """
   import Kernel, except: [send: 2]
 
   use GenServer